Numerous changes to support installation. CGI scripts are now based on ".xpls" template(1) sources which correctly build the header and allow the code to be aligned with current Antelope versioning [a strategy which should be repeated everywhere, to eliminate the unfortunate "current" link which is going to cause extensive difficulties at some point]. All web files are now install to a new WEBDIR target, which defaults to /tmp (since there's no predicting where the web directory will be on a given system) and can be customized with
        make install WEBDIR=/var/Web

If this is done with local mods to Makefile, please be careful not to submit
the local mod back to CVS.  The Makefile has been modified to support all of
the standard Antelope targets (install, uninstall, clean) for the web files.

A start was made on breaking out location-dependent values, like pathnames
for dot, the registry database, and the web tempfile directory into
orbtopo.pf. This file installs to the web directory rather than $ANTELOPE/data/pf.

Finally, a few small style changes were made to the dot graphs. Orb clients
are now blue boxes, to distinguish them from the yellow hexagons representing
the orbservers themselves. client pathnames are trimmed to be relative
instead of absolute, for clarity and propriety.

Remaining:
The symbol colors if not also the shapes should be extracted to the
parameter file as well at some point. More importantly, the xpls template
wisely adds -T for taint-checking. These scripts have not been cleaned up to
pass taint-checking, thus at the moment, after installation the -T flag
needs to be removed by hand in order to get them to run.