Multiple improvements to support movies in rt image bank. Multiple improvements to support movies in the real-time image bank.
I heavily restructured the parameter file, which now allows templates for
the straightforward creation of different output formats. The templates we
have figured out at the moment are for mpg (mpeg-1), mp4, and thumbnail
gifs. More can be added as we tackle the transcode codecs.  Support has
been added to move the generated movie files to sensible filenames with
the correct format extensions. I made this suffix generation automatic to
reduce the amount of labor in maintaining the parameter file.  The convert
command has been cleaned up with the use of a dereferenced list of file
names instead of having them all on the command line (and hence in the
log file, which was getting messy). Added a pf_revision_time check to
aid with the parameter-file upgrade.