Add option of automated backup of the archived collection to other physical resources on the SRB.
Add backup of the wfsrb table and the descriptor file to allow
future applications to access the waveforms via Datascope operations directly
through the SRB-Datascope driver.

Add tracking of the number of errors along with reporting at the end of
each run, and exit codes to correspond. [Notwithstanding an apparent bug
in Sbkupsrb(1) which returns 0 even when it fails]