initial checking for SRB_synch_orbregistries.
This
utility reads in orb information from Datascope (Antelope) registry
databases, which are pre-registered as SRB objects, and attempt to
build/synchronize a collection of all ORB data sources stored across all
databases. Each Orb source is registered as a SRB object, and can be
queried though SRB. SRB_synch_orbregistries also insert metadata for each SRB-ORB. It is very
useful if you want to query the ORBs based on this metadata.

Please read file "SETUP" to learn how to configure this program, and read the man file SRB_synch_orbregistries.1 to learn how to run the program.



-