A more robust version of SBE39 interface code.
Basically, there is no framing provided by the data logger for the data stream we are retrieving. I use a particular marker to denote the end of a data element. In the event that this marker does not arrive or is corrupted beyond recognition, this program can now time out.

In addition, there is some additional integrety checking for the data before it is sent to the ORB, this will eliminate unstuffing error messages.