(ice92orb)Release 1.30
Handle conversion of doubles off the wire, assuming sending and
receiving systems both use IEEE 754 for floating point representation
and that we can just swap the byte order around. Not sure if ARM is
going to play ball, so if this ever gets ported to a Marmot/Viper type
platform, it should be qualified there.