(ice92orb)Fix corrupted data, reflow text in file
There was an unnecessary conversion to network byte order occuring when
data was being read from the socket. Since htonl and friends are a no-op
on SPARC and other big-Endian architectures, we never tripped on this
bug.

Fixed some problems with my rewrites of the net/sta/chan/loc data
code to use constants - helps if I use the correct ones.

Fixed a missing byte order conversion, and cut down on the number of
ntoh* calls by caching values for data in the pkt structure.

Major reformatting of the code, which is increasing the diff count of
this commit even more.

This is release 1.31