Patch antelopemake on 5.3
patchF.0 for 5.3 updated Antelope to use a new licensing routine. This
changed a number of things, including the stock libraries.
Unfortunately, patched copies of include/antelopemake were not
distributed with this patch so compilation of various third party
programs fail.

This workaround needs to be removed once BRTT releases an official
patch.