Add the Python readline module
Per BRTT ticket 513, Antelope does not include the readline module for
licensing reasons. This leads to a severely hampered version of iPython,
and in some situations a broken interactive mode with the regular python
interpreter, especially with Python 2.7.5 and earlier on OSX 10.9
(Mavericks).

See
http://stackoverflow.com/questions/19531969/segmentation-fault-11-in-os-x
and http://bugs.python.org/issue18458 for background.