use tcp pings by default
Net::Ping needs root privs for ICMP ping. Many platforms have the ping
executable setuid root to work around this issue (not on Darwin though).

Incurring the sudo overhead isn't worth it for each probe, and we really
don't need the spam in our syslogs anyway. All of the TA stations
respond on 5380.