Switch all pings to tcp to avoid needing root
ICMP pings are unpredictable on cell modems unless the script is run as
root, even when using the Net::Ping::External module. Sidestep this
entirely by using the TCP probe option for all stations.

Switch this to use port 5381 - looks like 5380 doesn't respond even on
the wild blue stations where it is supposed to.

Make the matching logic a bit cleaner to read. Perl lacks a proper
switch or case statement, but this syntax should be a bit more obvious.