Consolidate insert and delete routines
Insert and deletes now use a dedicated data hash containing the list of
possible keys and values that can be output.

The insert and delete routines have an array that defines what fields
actually get output to the import file, rather than this being hardcoded
twice in slightly different fashions.