Move NOTIFY level down to priority 25
Juan's reading of the elog man page was incorrect. The Antelope Elog
Notify level is supposed to be informational, not an error.

Setting it to 25 should get it to print by default (which I think is
what he was going for), but not be so high as to trigger an elog warning
or critical notice.