Move anf and seispy to lib, add setup.py
The anf and seispy libraries never really belonged in data. Moving them to lib where they are easier to find for the uninitiated.

Fixed up various tests for the anf package.

Moved eloghandler under the anf build space.

Added setup.py files to pyanf and usarray_deploy_map to aid in
development. Not intended to actually install the files.